+++ Tests/SentryTests/SentryRequestTests.m
@@ -18,17 +18,9 @@ NSString *dsn = @"https://username:password@app.getsentry.com/12345";
 
 @interface SentryClient (Private)
 
-/**
- * Initializes a SentryClient which can be used for sending events to sentry.
- *
- * @param dsn DSN string of sentry
- * @param requestManager Object conforming SentryRequestManager protocol
- * @param error NSError reference object
- * @return SentryClient
- */
-- (_Nullable instancetype)initWithDsn:(NSString *)dsn
-                       requestManager:(id <SentryRequestManager>)requestManager
-                     didFailWithError:(NSError *_Nullable *_Nullable)error;
+- (_Nullable instancetype)initWithOptions:(NSDictionary<NSString *, id> *)options
+                           requestManager:(id <SentryRequestManager>)requestManager
+                         didFailWithError:(NSError *_Nullable *_Nullable)error;
 
 @end
 
@@ -181,18 +173,18 @@ NSString *dsn = @"https://username:password@app.getsentry.com/12345";
     [super setUp];
     [self clearAllFiles];
     self.requestManager = [[SentryMockRequestManager alloc] initWithSession:[SentryMockNSURLSession new]];
-    self.client = [[SentryClient alloc] initWithDsn:dsn
-                                     requestManager:self.requestManager
-                                   didFailWithError:nil];
+    self.client = [[SentryClient alloc] initWithOptions:@{@"dsn": dsn}
+                                         requestManager:self.requestManager
+                                       didFailWithError:nil];
     self.event = [[SentryEvent alloc] initWithLevel:kSentrySeverityDebug];
 }
-
+    
 - (void)testRealRequest {
     SentryQueueableRequestManager *requestManager = [[SentryQueueableRequestManager alloc] initWithSession:[SentryMockNSURLSession new]];
-    SentryClient *client = [[SentryClient alloc] initWithDsn:dsn
-                       requestManager:requestManager
-                     didFailWithError:nil];
-
+    SentryClient *client = [[SentryClient alloc] initWithOptions:@{@"dsn": dsn}
+                                                  requestManager:requestManager
+                                                didFailWithError:nil];
+    
     XCTestExpectation *expectation = [self expectationWithDescription:@"Request should finish"];
     [client sendEvent:self.event withCompletionHandler:^(NSError * _Nullable error) {
         XCTAssertNil(error);
@@ -256,9 +248,9 @@ NSString *dsn = @"https://username:password@app.getsentry.com/12345";
 
 - (void)testRequestQueueReady {
     SentryQueueableRequestManager *requestManager = [[SentryQueueableRequestManager alloc] initWithSession:[SentryMockNSURLSession new]];
-    SentryClient *client = [[SentryClient alloc] initWithDsn:dsn
-                                              requestManager:requestManager
-                                            didFailWithError:nil];
+    SentryClient *client = [[SentryClient alloc] initWithOptions:@{@"dsn": dsn}
+                                                  requestManager:requestManager
+                                                didFailWithError:nil];
 
     XCTAssertTrue(requestManager.isReady);
 
@@ -274,9 +266,9 @@ NSString *dsn = @"https://username:password@app.getsentry.com/12345";
 - (void)testRequestQueueCancel {
     SentryClient.logLevel = kSentryLogLevelVerbose;
     SentryQueueableRequestManager *requestManager = [[SentryQueueableRequestManager alloc] initWithSession:[SentryMockNSURLSession new]];
-    SentryClient *client = [[SentryClient alloc] initWithDsn:@"http://a:b@sentry.io/1"
-                                              requestManager:requestManager
-                                            didFailWithError:nil];
+    SentryClient *client = [[SentryClient alloc] initWithOptions:@{@"dsn": @"http://a:b@sentry.io/1"}
+                                                  requestManager:requestManager
+                                                didFailWithError:nil];
 
     XCTestExpectation *expectation = [self expectationWithDescription:@"Request should fail"];
     [client sendEvent:self.event withCompletionHandler:^(NSError * _Nullable error) {
@@ -831,4 +823,17 @@ NSString *dsn = @"https://username:password@app.getsentry.com/12345";
     XCTAssertEqual([fileManager getAllStoredEvents].count, (unsigned long)0);
 }
 
+- (void)testDisabledClient {
+    NSError *error = nil;
+    SentryFileManager *fileManager = [[SentryFileManager alloc] initWithDsn:[[SentryDsn alloc] initWithString:dsn didFailWithError:nil] didFailWithError:&error];
+    
+    SentryEvent *event = [[SentryEvent alloc] initWithLevel:kSentrySeverityWarning];
+    event.message = @"abc";
+    SentryClient.logLevel = kSentryLogLevelDebug;
+    self.client.enabled = @NO;
+    [self.client sendEvent:event withCompletionHandler:nil];
+    
+    XCTAssertEqual([fileManager getAllStoredEvents].count, (unsigned long)1);
+}
+    
 @end